project-front/django-user-restapi/pyproject.toml
[tool.poetry]
name = "django-api-demo"
version = "0.1.0"
description = ""
authors = ["LuxProtoss <zerox12311@gmail.comm>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
coreapi = "^2.3.3"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.6.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
